lifts up standings state to app for use across app
